[Design] 검색페이지 컴포넌트 분리 (#93)
